Expand exploration of format to make sure I'm not missing anything else base on case sensitivity or leading whitespace.
For reference, at the moment the final version from his blog runs in ~1m20 on my machine
while the naive perl version runs in ~2m20. Not sure if the slower time is due to reading
more files or that I'm running in a local virtual machine on windows (which generally has
disk read times about 1/3 of linux for small files from what I remember).